* 'main' of https://github.com/InCourseProject/in-course-be:
  찜하기 기능 수정
  코스/ 장소 찜하기 controller, service 분리
  코스(게시글) 찜하기, 스코어 기능 관련 코드 수정
  코슥(게시글) 찜하기 기능 코드 리팩토링
  장소(카드)찜하기 기능 코드 리팩토링
  member의 닉네임 컬럼 notnull로 변경(회원 가입 시 랜던값이 들어가고, 회원가입 세부사항에서 닉네임 변경)
  스코어 중복 검사 추가(이미 점수를 줬으면 기존 점수 삭제 후 점수 저장)
  찜하기 코스, 찜하기 장소 분리해서 관리, 기존 찜하기 삭제

+++ src/main/java/com/example/week08/dto/response/PostResponseGetDto.java
@@ -21,7 +21,7 @@ public class PostResponseGetDto {
     private String region;
     private String season;
     private String who;
-    private int avgScore;
+    private double avgScore;
     private int heart;
     private LocalDateTime createdAt;
     private LocalDateTime modifiedAt;
